requirements.txt
#for streamlit
streamlit==1.2.0
matplotlib

#for Deep learning model
tensorflow==2.5.0
scikit-learn==1.0.2

#Image
Pillow==9.0.0
scikit-image==0.19.1

#other
numpy==1.19.5
pandas==1.3.5

joblib==1.1.0
